密码加密、用户退出代码优化

+++ src/layout/header/LoginOut.vue
@@ -33,7 +33,7 @@ import useInstance from "@/hooks/useInstance";
 import { userStore } from "@/store/user";
 import { ElMessage, FormInstance } from "element-plus";
 import { reactive, ref } from "vue";
-import { updatePasswordApi } from "@/api/home";
+import { updatePasswordApi, loginOutApi } from "@/api/home";
 const { dialog, onClose, onConfirm, onShow } = useDialog();
 const { global } = useInstance();
 const upForm = ref<FormInstance>();
@@ -41,12 +41,15 @@ const store = userStore();
 const loginOut = async () => {
     const confirm = await global.$myconfirm("确定退出登录吗？");
     if (confirm) {
-        //清空用户信息
-        store.setToken("");
-        store.setUserId("");
-        store.setUserType("");
-        localStorage.clear();
-        window.location.href = "/login";
+        let res = await loginOutApi()
+        if(res && res.code == 200){
+            //清空用户信息
+            store.setToken("");
+            store.setUserId("");
+            store.setUserType("");
+            localStorage.clear();
+            window.location.href = "/login";
+        }
     }
 };
 const updateBtn = () => {
